pyproject.toml
[project]
name = "langgraph-advanced-tutorial"
version = "0.1.0"
description = "LangGraph Advanced Tutorial"
readme = "README.md"
authors = [
    { name="Teddy", email="teddylee777@gmail.com" },
    { name="Hyunjun Jeon", email="jeonhj920@gmail.com" },
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "a2a-sdk[all]>=0.3.11",
    "arxiv>=2.3.0",
    "chonkie>=1.4.1",
    "deepagents>=0.2.5",
    "deepeval>=3.7.0",
    "docling>=2.61.2",
    "faiss-cpu>=1.12.0",
    "jupyter>=1.1.1",
    "langchain>=1.0.2",
    "langchain-anthropic>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.1.12",
    "langchain-neo4j>=0.6.0",
    "langchain-openai>=1.0.1",
    "langchain-qdrant>=1.1.0",
    "langchain-tavily>=0.2.12",
    "langchain-teddynote>=0.5.4",
    "langfuse>=3.9.3",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-checkpoint-redis>=0.1.2",
    "langgraph-cli[inmem]>=0.4.7",
    "llama-parse>=0.6.79",
    "matplotlib>=3.10.7",
    "mcp>=1.21.0",
    "mem0ai>=1.0.0",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
    "pymupdf>=1.26.6",
    "pymupdf4llm>=0.2.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.3.8",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "tiktoken>=0.12.0",
]
